Calculate visible region two different ways
Since Windows 8+ insists on behaving badly with respect to calculating
the visible region by taking the window size and adjusting by the size
of scrollbars, find an alternate solution that seems to work on Windows.
Compensate for problems previously found on Linux by continuing to
calculate the "old" way, but take the minimum of the two methods.

(Hopefully) resolves #556